chore: bind nvm to whole circle ci
Each job in circle ci is considered as like different terminal tabs. Thus the defualt node version set by nvm in a job
does not impact globally. Previously it reste the NVM env var on demand. Now set it to BASH_ENV to allow every job to pick
it up.